フェッチ処理をカスタムフックにした
